+++ integrations/fastify/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @scalar/fastify-api-reference
 
+## 1.59.3
+
 ## 1.59.2
 
 ## 1.59.1