+++ packages/api-client-react/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @scalar/api-client-react
 
+## 2.0.25
+
 ## 2.0.24
 
 ## 2.0.23